+++ app/Http/Controllers/Laporan_bulananController.php
@@ -27,23 +27,14 @@ class Laporan_bulananController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        $laporan_bulanan = Laporan_bulanan::create($request->all());
-        // return redirect()->route("laporan_bulanan.index")->with("success","");
-        return response()->json($laporan_bulanan);
-    }
-
-    /**
      * Display the specified resource.
      */
     public function show(string $id)
     {
         $laporan_bulanan = Laporan_bulanan::find($id);
+        if (!$laporan_bulanan) return response()->json(null,404);
         // return view("laporan_bulanan.show", compact("laporan_bulanan"));
-        return response()->json(($laporan_bulanan) ? $laporan_bulanan : null,404);
+        return response()->json($laporan_bulanan);
     }
 
     /**
@@ -52,8 +43,9 @@ class Laporan_bulananController extends Controller
     public function edit(string $id)
     {
         $laporan_bulanan = Laporan_bulanan::find($id);
+        if (!$laporan_bulanan) return response()->json(null,404);
         // return view("laporan_bulanan.edit", compact("laporan_bulanan"));
-        return response()->json(($laporan_bulanan) ? $laporan_bulanan : null,404);
+        return response()->json($laporan_bulanan);
     }
 
     /**
@@ -62,9 +54,10 @@ class Laporan_bulananController extends Controller
     public function update(Request $request, string $id)
     {
         $laporan_bulanan = Laporan_bulanan::find($id);
+        if (!$laporan_bulanan) return response()->json(null,404);
         $laporan_bulanan->update($request->all());
         // return redirect()->route("laporan_bulanan.index")->with("success","");
-        return response()->json(($laporan_bulanan) ? $laporan_bulanan : null,404);
+        return response()->json($laporan_bulanan)->with("success","");
     }
 
     /**
@@ -73,8 +66,9 @@ class Laporan_bulananController extends Controller
     public function destroy(string $id)
     {
         $laporan_bulanan = Laporan_bulanan::find($id);
+        if (!$laporan_bulanan) return response()->json(null,404);
         $laporan_bulanan->delete();
         // return redirect()->route("laporan_bulanan.index")->with("success","");
-        return response()->json(($laporan_bulanan) ? $laporan_bulanan : null,404);
+        return response()->json(null)->with("success","");
     }
 }
